move db to lib and fix drizzle setup
