+++ drizzle.config.ts
@@ -1,8 +1,13 @@
+import "dotenv/config";
 import { defineConfig } from "drizzle-kit";
 
+if (!process.env.DATABASE_URL) {
+  throw new Error("DATABASE_URL environment variable is not set");
+}
+
 export default defineConfig({
-  schema: "./db/schema.ts",
-  out: "./drizzle",
+  schema: "./lib/db/schema.ts",
+  out: "./lib/db/migrations",
   dialect: "postgresql",
   dbCredentials: {
     url: process.env.DATABASE_URL ?? "",
